Fix text tool broken by #1543 (#1555)
Text tool setting font as f32

+++ editor/src/messages/portfolio/document/node_graph/node_graph_message_handler/document_node_types.rs
@@ -2483,7 +2483,7 @@ fn static_nodes() -> Vec<DocumentNodeDefinition> {
 							NodeInput::Network(concrete!(application_io::EditorApi<graphene_std::wasm_application_io::WasmApplicationIo>)),
 							NodeInput::Network(concrete!(String)),
 							NodeInput::Network(concrete!(graphene_core::text::Font)),
-							NodeInput::Network(concrete!(f64)),
+							NodeInput::Network(concrete!(f32)),
 						],
 						implementation: DocumentNodeImplementation::Unresolved(ProtoNodeIdentifier::new("graphene_core::text::TextGeneratorNode<_, _, _>")),
 						..Default::default()
